SAP Project/requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-RESTful==0.3.10
Flask-JWT-Extended==4.5.2
Flask-Cors==4.0.0
SQLAlchemy==2.0.20
bcrypt==4.0.1
pandas==2.1.0
scikit-learn==1.3.0
requests==2.31.0
python-dotenv==1.0.0
gunicorn==21.2.0
